Create HasScoreStats trait for Scorable models.
ScoreStats in Results page is collapsible.

Remove DialogActions from QuizzerWindow.

Replace hard-coded `deck` type with `scorable_type`.
