+++ app/Models/Deck.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use App\Models\Scopes\DeckScope;
+use App\Models\Traits\HasScoreStats;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -13,6 +14,7 @@ use Maize\Markable\Models\Bookmark;
 
 class Deck extends Model
 {
+    use HasScoreStats;
     use HasFactory;
     use Markable;
 
@@ -68,61 +70,6 @@ class Deck extends Model
     public function scores(): MorphMany
     {
         return $this->morphMany(Score::class, 'scorable')->orderByDesc('created_at');
-    }
-
-    public function getScoreStatsAttribute()
-    {
-        $scores = $this->scores;
-
-        $latest = $scores->first();
-        $highest = $scores->sortByDesc('score')->first();
-        $average = $scores->avg('score');
-        $count = $scores->count();
-
-        $latestTrend = null;
-        $highestTrend = null;
-        $averageTrend = null;
-
-        if ($latest) {
-            $previous = $scores->skip(1)->first();
-
-            if ($previous) {
-                $latestTrend = $latest->score > $previous->score ? 'up' : 'down';
-                if ($latest->score === $previous->score) $latestTrend = null;
-
-            } else {
-                $latestTrend = 'up';
-            }
-
-            if ($scores->count() > 1) {
-                $allExceptLatest = $scores->slice(1);
-
-                $highestWithoutLatest = $allExceptLatest->max('score');
-
-                if ($latest->score > $highestWithoutLatest) $highestTrend = 'up';
-
-                $averageWithoutLatest = $allExceptLatest->avg('score');
-
-                $averageTrend = $latest->score > $averageWithoutLatest  ? 'up' : 'down';
-                if ($latest->score === $averageWithoutLatest) $averageTrend = null;
-
-            } else {
-                $highestTrend = 'up';
-                $averageTrend = 'up';
-            }
-        }
-
-        return [
-            'latest' => $latest?->score,
-            'latest_date' => $latest?->created_at?->format('j F Y'),
-            'latest_trend' => $latestTrend,
-            'highest' => $highest?->score,
-            'highest_date' => $highest?->created_at?->format('j F Y'),
-            'highest_trend' => $highestTrend,
-            'average' => round($average, 2),
-            'average_trend' => $averageTrend,
-            'count' => $count,
-        ];
     }
 
     public function scopeFilter($query, array $filters): void
